refactor: move inside ui
