+++ src/components/TopBar/TopBar.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 import { ReactNode } from 'react';
 import styled from 'styled-components';
-import MaxWidthWrapper from 'src/components/ui/MaxWidthWrapper';
+import MaxWidthWrapper from 'src/components/MaxWidthWrapper';
 
 export default function TopBar({ children }: { children: ReactNode }) {
   return (
