+++ src/pages/index.tsx
@@ -3,9 +3,9 @@ import Head from 'next/head';
 import styled from 'styled-components';
 import { microcms } from '@/libs/microcms';
 import { Blog, Post } from '@/types/blog';
+import MaxWidthWrapper from 'src/components/MaxWidthWrapper';
 import Card from 'src/components/home/Card';
 import { Pagination } from 'src/components/home/Pagination';
-import MaxWidthWrapper from 'src/components/ui/MaxWidthWrapper';
 
 type Props = {
   blog: Post[];
